# docs/swagger.yaml
definitions:
...
  github_com_ThalysSilva_unicast-backend_internal_user.User:
    properties:
      email:
        type: string
      id:
        type: string
      name:
        type: string
    required:
    - email
    - name
    type: object
  github_com_ThalysSilva_unicast-backend_pkg_api.DefaultResponse-array_internal_campus_Campus:
    properties:
      data:
        items:
          $ref: '#/definitions/internal_campus.Campus'
        type: array
      message:
        type: string
    type: object
  github_com_ThalysSilva_unicast-backend_pkg_api.DefaultResponse-array_internal_course_Course:
    properties:
      data:
        items:
          $ref: '#/definitions/internal_course.Course'
        type: array
      message:
        type: string
    type: object
...
  github_com_ThalysSilva_unicast-backend_pkg_api.DefaultResponse-internal_auth_LoginResponse:
    properties:
      data:
        $ref: '#/definitions/internal_auth.LoginResponse'
      message:
        type: string
    type: object
  github_com_ThalysSilva_unicast-backend_pkg_api.DefaultResponse-internal_auth_RefreshResponse:
    properties:
      data:
        $ref: '#/definitions/internal_auth.RefreshResponse'
      message:
        type: string
    type: object
...
  github_com_ThalysSilva_unicast-backend_pkg_api.DefaultResponse-map_string_string:
    properties:
      data:
        $ref: '#/definitions/map_string_string'
      message:
        type: string
    type: object
  github_com_ThalysSilva_unicast-backend_pkg_api.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  internal_auth.LoginInput:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  internal_auth.LoginResponse:
    properties:
      accessToken:
        type: string
      jwe:
        type: string
      refreshToken:
        type: string
      user:
        $ref: '#/definitions/github_com_ThalysSilva_unicast-backend_internal_user.User'
    type: object
  internal_auth.RefreshInput:
    properties:
      refreshToken:
        type: string
    required:
    - refreshToken
    type: object
  internal_auth.RefreshResponse:
    properties:
      accessToken:
        type: string
      refreshToken:
        type: string
      user:
        $ref: '#/definitions/github_com_ThalysSilva_unicast-backend_internal_user.User'
    type: object
  internal_auth.RegisterInput:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    required:
    - email
    - name
    - password
    type: object
  internal_backdoor.resetInput:
    properties:
      email:
        type: string
      newPassword:
        type: string
      secret:
        type: string
      userId:
        type: string
    required:
    - newPassword
    - secret
    type: object
  internal_campus.Campus:
    properties:
      description:
        type: string
      id:
        type: string
      name:
        type: string
    required:
    - name
    type: object
  internal_campus.createCampusInput:
    properties:
      description:
        type: string
      name:
        type: string
    required:
    - description
    - name
    type: object
  internal_campus.updateCampusInput:
    properties:
      description:
        type: string
      name:
        type: string
    type: object
  internal_course.Course:
    properties:
      description:
        type: string
      id:
        type: string
      name:
        type: string
      semester:
        type: integer
      year:
        type: integer
    required:
    - name
    - semester
    - year
    type: object
  internal_course.createCourseInput:
    properties:
      description:
        type: string
      name:
        type: string
      semester:
        type: integer
      year:
        type: integer
    required:
    - description
    - name
    - semester
    - year
    type: object
  internal_course.updateCourseInput:
    properties:
      description:
        type: string
      name:
        type: string
      semester:
        type: integer
      year:
        type: integer
    type: object
...
  map_string_string:
    additionalProperties:
      type: string
    type: object
info:
  contact: {}
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Gera o acesso a um usuário no sistema
      parameters:
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/internal_auth.LoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ThalysSilva_unicast-backend_pkg_api.DefaultResponse-internal_auth_LoginResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_ThalysSilva_unicast-backend_pkg_api.ErrorResponse'
      summary: Gera o acesso a um usuário
      tags:
      - auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Remove o acesso a um usuário do sistema
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              message:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_ThalysSilva_unicast-backend_pkg_api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Remove o acesso a um usuário
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Atualiza o Refresh Token do usuário no sistema
      parameters:
      - description: Refresh token
        in: body
        name: refreshToken
        required: true
        schema:
          $ref: '#/definitions/internal_auth.RefreshInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ThalysSilva_unicast-backend_pkg_api.DefaultResponse-internal_auth_RefreshResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_ThalysSilva_unicast-backend_pkg_api.ErrorResponse'
      summary: Atualiza o Refresh Token do usuário
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Registra um novo usuário no sistema
      parameters:
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/internal_auth.RegisterInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_ThalysSilva_unicast-backend_pkg_api.ErrorResponse'
      summary: Registra um novo usuário
      tags:
      - auth
  /backdoor/reset-password:
    post:
      consumes:
      - application/json
      description: Uso administrativo com segredo estático; requer userId ou email.
      parameters:
      - description: Dados de reset
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/internal_backdoor.resetInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ThalysSilva_unicast-backend_pkg_api.DefaultResponse-map_string_string'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_ThalysSilva_unicast-backend_pkg_api.ErrorResponse'
      summary: Reseta a senha de um usuário (backdoor)
      tags:
      - backdoor
  /campus:
    get:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ThalysSilva_unicast-backend_pkg_api.DefaultResponse-array_internal_campus_Campus'
      summary: Lista campi
      tags:
      - campus
    post:
      consumes:
      - application/json
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Dados do campus
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_campus.createCampusInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ThalysSilva_unicast-backend_pkg_api.DefaultResponse-map_string_string'
      summary: Cria um campus
      tags:
      - campus
  /campus/{id}:
    put:
      consumes:
      - application/json
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Campus ID
        in: path
        name: id
        required: true
        type: string
      - description: Campos para atualizar
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_campus.updateCampusInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ThalysSilva_unicast-backend_pkg_api.DefaultResponse-map_string_string'
      summary: Atualiza um campus
      tags:
      - campus
  /course:
    post:
      consumes:
      - application/json
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Dados da disciplina
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_course.createCourseInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ThalysSilva_unicast-backend_pkg_api.DefaultResponse-map_string_string'
      summary: Cria uma disciplina
      tags:
      - course
...
  /course/{id}:
    delete:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Course ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ThalysSilva_unicast-backend_pkg_api.DefaultResponse-map_string_string'
      summary: Deleta uma disciplina
      tags:
      - course
    put:
      consumes:
      - application/json
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Course ID
        in: path
        name: id
        required: true
        type: string
      - description: Campos para atualizar
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_course.updateCourseInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ThalysSilva_unicast-backend_pkg_api.DefaultResponse-map_string_string'
      summary: Atualiza uma disciplina
      tags:
      - course
  /course/{programId}:
    get:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ThalysSilva_unicast-backend_pkg_api.DefaultResponse-array_internal_course_Course'
      summary: Lista disciplinas do usuário
      tags:
      - course
...
swagger: "2.0"